Prog 494
Must go after the following DB RFCs: 1) https://jira.corp.inmobi.com/browse/DBA-3062 2) https://jira.corp.inmobi.com/browse/DBA-3063  Design Doc: https://docs.google.com/document/d/10esGdE-DT3QYzR5uJKM4VabvXDU2vW-8IPSAXMyerrg/edit#  New Unit Tests count: 40  Misc Refactoring: Changed SasParams.modelId to Long from Int. (It is a long in the AdPoolRequest)=ishan-bhatnagar

+++ Server/src/main/java/com/inmobi/adserve/channels/server/requesthandler/filters/adgroup/impl/AdGroupHandsetTargetingFilter.java
@@ -38,7 +38,7 @@ public class AdGroupHandsetTargetingFilter extends AbstractAdGroupLevelFilter {
     protected boolean failedInFilter(final ChannelSegment channelSegment, final SASRequestParameters sasParams,
             final CasContext casContext) {
 
-        final List<Integer> manufModelTargetingList =
+        final List<Long> manufModelTargetingList =
                 channelSegment.getChannelSegmentEntity().getManufModelTargetingList();
         return CollectionUtils.isNotEmpty(manufModelTargetingList)
                 && !manufModelTargetingList.contains(sasParams.getModelId());
